+++ app/controllers/project_memberships_controller.rb
@@ -22,7 +22,7 @@ class ProjectMembershipsController < ApplicationController
             turbo_stream.replace("project_membership_message", partial: "project_memberships/success_message", locals: { email: user.email })
           ]
         end
-        format.html { redirect_to settings_project_path(@project), notice: "Project shared with #{user.email}." }
+        format.html { redirect_to settings_project_path(@project, section: "team"), notice: "Project shared with #{user.email}." }
       end
     else
       respond_with_membership_error(membership.errors.full_messages.to_sentence)
@@ -46,7 +46,7 @@ class ProjectMembershipsController < ApplicationController
           )
         ]
       end
-      format.html { redirect_to settings_project_path(@project), notice: "Access removed." }
+      format.html { redirect_to settings_project_path(@project, section: "team"), notice: "Access removed." }
     end
   end
 
@@ -61,7 +61,7 @@ class ProjectMembershipsController < ApplicationController
   end
 
   def redirect_with_alert(message)
-    redirect_to settings_project_path(@project), alert: message
+    redirect_to settings_project_path(@project, section: "team"), alert: message
   end
 
   def respond_with_membership_error(message)
@@ -71,7 +71,7 @@ class ProjectMembershipsController < ApplicationController
           partial: "project_memberships/error_message",
           locals: { message: message }), status: :unprocessable_content
       end
-      format.html { redirect_to settings_project_path(@project), alert: message }
+      format.html { redirect_to settings_project_path(@project, section: "team"), alert: message }
     end
   end
 end